move content page to course-overview
