+++ src/app/account/dashboard/explore/components/course.tsx
@@ -2,10 +2,10 @@ import { Bookmark, Clock4 } from "lucide-react";
 import img from "@/public/img/Rectangle 8(1).svg";
 import Image from "next/image";
 
-function CourseCard(){
+function CourseCard({className}: {className?: string}) {
     return (
-      <div className="border max-w-96 w-[352px] rounded-lg border-[#313130]">
-        <Image src={img} alt="course logo" />
+      <div className={`border w-[352px] rounded-lg border-[#313130] ${className}`}>
+        <Image src={img} alt="course logo" className="w-full"/>
         <div className="py-6 px-2">
           <div className="grid grid-cols-[3fr_1fr]">
             <h2 className="text-xl font-medium text-[#FCFCFC]">
